perf(cli): keep urllib.request and install.py off the startup path
utils imported comfy_cli.http for one timeout constant; comfy_cli.http
imports urllib.request, about 60 ms, and utils is on every invocation's
path. Import the constant inside the one download helper that uses it.

cmdline's install and update commands use validate_version and
validate_optional_version as typer callbacks, which typer needs at
decoration time, so their module is imported at startup. Their previous
home, command/install.py, imports requests and the custom-nodes machinery.
Move the two functions to command/version_validators.py; install.py
re-exports them so existing imports and tests are unchanged.
